@@ -4,13 +4,19 @@ import { LoginComponent } from './components/login/login.component';
 import { RegisterComponent } from './components/register/register.component';
 import { ManagementPanelComponent } from './components/management-panel/management-panel/management-panel.component';
 import { DashboardComponent } from './components/management-panel/dashboard/dashboard/dashboard.component';
+import { ProjectsComponent } from './components/management-panel/projects/projects.component';
+import { AddProjectComponent } from './components/management-panel/projects/add-project/add-project.component';
+import { ProjectDetailsComponent } from '@components/management-panel/dashboard/project-details/project-details.component';
+import { AddTeamComponent } from '@components/management-panel/projects/add-team/add-team.component';
+import { AddProjectPartComponent } from '@components/management-panel/projects/add-project-part/add-project-part.component';
+import { AuthGuard } from '@shared/auth.guard';
 
 const routes: Routes = [
-  {
+  /*{
     path: '',
     redirectTo: '/login',
     pathMatch: 'full'
-  },
+  },*/
   {
     path: 'login',
     component: LoginComponent
@@ -20,17 +26,40 @@ const routes: Routes = [
     component: RegisterComponent
   },
   {
-    path: 'panel',
+    path: '',
+    canActivate: [
+        AuthGuard
+    ],
     component: ManagementPanelComponent,
     children: [
       {
         path: '',
         redirectTo: 'dashboard',
         pathMatch: 'full'
-      },  
+      },
       {
         path: 'dashboard',
         component: DashboardComponent
+      },
+      {
+        path: 'projects',
+        component: ProjectsComponent
+      },
+      {
+        path: 'project/add',
+        component: AddProjectComponent
+      },
+      {
+        path: 'project/:id',
+        component: ProjectDetailsComponent
+      },
+      {
+          path: 'project/:id/teams/add',
+          component: AddTeamComponent
+      },
+      {
+        path: 'project/:id/parts/add',
+        component: AddProjectPartComponent
       }
     ]
   }
