@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { LoaderService } from 'src/app/providers/loader/loader.service';
+import { LoaderService, UserService } from '@providers';
+import { LoginResponse } from '@shared';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-login',
@@ -11,27 +13,43 @@ export class LoginComponent implements OnInit {
   email: string = '';
   password: string = '';
 
-  constructor(private loader: LoaderService){
-
+constructor(private _loader: LoaderService,
+  private _user: UserService,
+  private _router: Router) {
   }
 
-  ngOnInit(): void{
-        
+  ngOnInit(): void {
   }
 
-  onLogin(): void{
+  onLogin(): void {
     this.makeLoginRequest();
   }
 
-  onKeyDown(): void{
+  // why does this function exist??? @elitru?
+  onKeyDown(): void {
     this.makeLoginRequest();
   }
 
-  private makeLoginRequest(): void{
-    if(this.email === '' || this.password === ''){
+  private makeLoginRequest(): void {
+    if (!this.email || !this.password) {
       return;
     }
 
-    this.loader.setVisible(true);
+    this._loader.setVisible(true);
+
+    // login part
+    const s = this._user.login({email: this.email, password: this.password}).subscribe((res: LoginResponse) => {
+      s.unsubscribe();
+      this._user.user.next(res.user);
+      this._user.token.next(res.token);
+      this._user.tokenExpiration.next(new Date(res.expiration));
+      this._user.save();
+      this._loader.setVisible(false);
+      this._router.navigateByUrl("/dashboard");
+    }, err => {
+      // TODO add dialog
+      console.error(err);
+    });
+
   }
 }
